Bug fixing para obtener oferta ganadora y vistas refactorizadas completas

+++ trunk/src/edu/tdp2/client/model/Oferta.java
@@ -183,7 +183,7 @@ public class Oferta extends AbstractDomainObject
 				return false;
 			if (!o.getMoneda().equals(getMoneda()))
 				return false;
-			if (o.getUsuario().getId() != getUsuario().getId())
+			if (o.getUsuario().getId().longValue() != getUsuario().getId().longValue())
 				return false;
 			
 			return true;
